requirements.txt
python-telegram-bot==20.7
yt-dlp==2024.4.9
pydub==0.25.1
moviepy==1.0.3
requests==2.31.0